pyproject.toml
[project]
name = "workflows-with-airflow"
version = "0.1.0"
description = "Academy course for workflows with Apache Airflow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow>=3.1.6,<3.2.0",
    "apache-airflow-providers-standard",
    "pandas>=2.2.3",
]
